requirements.txt
kafka-python==2.2.15
psutil==7.0.0
